libs/lib.py
import torch
import torch.nn as nn
from torch.utils.data import Dataset, DataLoader
import torch.optim as optim
import numpy as np 
from PIL import Image
import matplotlib.pyplot as plt
import torchvision
from torchvision import datasets, models, transforms
import os
from sklearn.preprocessing import LabelEncoder, OneHotEncoder
from glob import glob
import json
import time
import tqdm