requirements.txt
torch
torchvision
torchaudio
matplotlib
numpy
sklearn
Pillow
copy
tqdm
shutil
argparse